LayoutTests/media/video-large-pixel-aspect-ratio-expected.txt

Test that a video with a large pixel-aspect-ratio has the correct display width and display height.
EXPECTED (video.videoWidth == '331') OK
EXPECTED (video.videoHeight == '196') OK
END OF TEST
